ml_lifecycle/requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.66.0
h11==0.12.0
idna==3.2
importlib-metadata==3.10.1
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pymongo==3.11.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
starlette==0.14.2
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.5.0
